+++ test/src/run_script.dart
@@ -1,3 +1,4 @@
+// @dart=2.9
 // Copyright 2013 Google. All rights reserved. Use of this source code is
 // governed by a BSD-style license that can be found in the LICENSE file.