+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # grinder.dart changes
 
+## 0.8.7
+- Declare `Never` instead of `Null` for `fail()` and `GrinderContext.fail()`.
+- Make the `args` parameter to `GrinderTask.execute()` optional.
+- Support null-safety mode.
+
 ## 0.8.6
 - Declare a `Null` return type for `fail()` and `GrinderContext.fail()`.